+++ test/commands/issue/issue-view.test.ts
@@ -7,7 +7,7 @@ const TEST_ENDPOINT = "http://127.0.0.1:59123/graphql"
 
 // Common Deno args for permissions
 const denoArgs = [
-  "--allow-env=GITHUB_*,GH_*,LINEAR_*,NODE_ENV,EDITOR,SNAPSHOT_TEST_NAME",
+  "--allow-env=GITHUB_*,GH_*,LINEAR_*,NODE_ENV,EDITOR,PAGER,SNAPSHOT_TEST_NAME,CLIFFY_SNAPSHOT_FAKE_TIME,NO_COLOR,TMPDIR,TMP,TEMP",
   "--allow-read",
   "--allow-write",
   "--allow-run",